# Conflicts:
#	EduTech/EduTech/src/main/resources/application.properties
#	EduTech/EduTech/target/classes/application.properties

+++ EduTech/EduTech/src/main/java/com/EduTech/EduTech/controller/CursoController.java
@@ -61,7 +61,7 @@ public class    CursoController {
         }
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable Long id){
+    public ResponseEntity<?> eliminar(@PathVariable Integer id){
         try{
             cursoService.delete(id);
             return ResponseEntity.noContent().build();
